+++ homeassistant/components/http/const.py
@@ -1,0 +1,12 @@
+"""HTTP specific constants."""
+KEY_AUTHENTICATED = 'ha_authenticated'
+KEY_USE_X_FORWARDED_FOR = 'ha_use_x_forwarded_for'
+KEY_TRUSTED_NETWORKS = 'ha_trusted_networks'
+KEY_REAL_IP = 'ha_real_ip'
+KEY_BANS_ENABLED = 'ha_bans_enabled'
+KEY_BANNED_IPS = 'ha_banned_ips'
+KEY_FAILED_LOGIN_ATTEMPTS = 'ha_failed_login_attempts'
+KEY_LOGIN_THRESHOLD = 'ha_login_treshold'
+KEY_DEVELOPMENT = 'ha_development'
+
+HTTP_HEADER_X_FORWARDED_FOR = 'X-Forwarded-For'
