+++ homeassistant/components/ios.py
@@ -250,11 +250,10 @@ def setup(hass, config):
 
     discovery.load_platform(hass, "sensor", DOMAIN, {}, config)
 
-    hass.http.register_view(iOSIdentifyDeviceView(hass))
+    hass.http.register_view(iOSIdentifyDeviceView)
 
     app_config = config.get(DOMAIN, {})
-    hass.http.register_view(iOSPushConfigView(hass,
-                                              app_config.get(CONF_PUSH, {})))
+    hass.http.register_view(iOSPushConfigView(app_config.get(CONF_PUSH, {})))
 
     return True
 
@@ -266,9 +265,8 @@ class iOSPushConfigView(HomeAssistantView):
     url = "/api/ios/push"
     name = "api:ios:push"
 
-    def __init__(self, hass, push_config):
+    def __init__(self, push_config):
         """Init the view."""
-        super().__init__(hass)
         self.push_config = push_config
 
     @callback
@@ -282,10 +280,6 @@ class iOSIdentifyDeviceView(HomeAssistantView):
 
     url = "/api/ios/identify"
     name = "api:ios:identify"
-
-    def __init__(self, hass):
-        """Init the view."""
-        super().__init__(hass)
 
     @asyncio.coroutine
     def post(self, request):
