+++ homeassistant/components/switch/netio.py
@@ -97,6 +97,7 @@ class NetioApiView(HomeAssistantView):
     @callback
     def get(self, request, host):
         """Request handler."""
+        hass = request.app['hass']
         data = request.GET
         states, consumptions, cumulated_consumptions, start_dates = \
             [], [], [], []
@@ -119,7 +120,7 @@ class NetioApiView(HomeAssistantView):
         ndev.start_dates = start_dates
 
         for dev in DEVICES[host].entities:
-            self.hass.async_add_job(dev.async_update_ha_state())
+            hass.async_add_job(dev.async_update_ha_state())
 
         return self.json(True)
